Travel Loader/requirements.txt
spacy
sounddevice
google.api_core.exceptions
scipy.io.wavfile
networkx
matplotlib
pyplot